requirements.txt
# === Agent + LLM Tools ===
# aisuite==0.1.11  # Package not found on PyPI - install manually if needed
anthropic
docstring-parser
markdown
mistralai
openai
qrcode
tavily-python>=0.7.12
textstat
google-cloud-aiplatform  # vertexai replacement

# === Web Framework + API ===
fastapi
pydantic
pydantic[email]
python-dotenv
python-multipart
requests
sqlalchemy
uvicorn

# === Notebook Experience ===
ipywidgets
jupyter_server
nbclassic
notebook

# === Data Analysis / Display (Optional Enhancements) ===
# duckdb  # Removed due to zstd library compilation errors - install manually if needed
matplotlib
pandas
seaborn
tabulate
tinydb

# === Architecture Diagrams ===
diagrams>=0.23.0

# === Machine Learning / NLP (Optional Enhancements) ===
jinja2
psycopg2-binary
scikit-learn
# wikipedia-api  # Removed due to setuptools issues - install manually if needed 